client/driver: gaurd authHelper test from running on windows
